Get rid of helpers.go
