+++ fs.go
@@ -8,34 +8,13 @@ type DirMaker interface {
 	MkdirAll(path string, perm os.FileMode) error
 }
 
-type FakeDirMaker struct {
-	MkdirAllError error
-}
-
-func (f *FakeDirMaker) MkdirAll(path string, perm os.FileMode) error {
-	return f.MkdirAllError
-}
-
 type Symlinker interface {
 	Symlink(oldname, newname string) error
-}
-
-type FakeSymlinker struct {
-	SymlinkError error
-}
-
-func (f *FakeSymlinker) Symlink(oldname, newname string) error {
-	return f.SymlinkError
 }
 
 type Stater interface {
 	Stat(name string) (os.FileInfo, error)
 	IsNotExist(err error) bool
-}
-
-type MkdirSymlinker interface {
-	DirMaker
-	Symlinker
 }
 
 type FakeMkdirSymlinker struct {
@@ -48,6 +27,43 @@ type StatDirMaker interface {
 	DirMaker
 }
 
+// Fake implementation of interface
+
+type FakeDirMaker struct {
+	MkdirAllError error
+}
+
+func (f *FakeDirMaker) MkdirAll(path string, perm os.FileMode) error {
+	return f.MkdirAllError
+}
+
+type FakeSymlinker struct {
+	SymlinkError error
+}
+
+func (f *FakeSymlinker) Symlink(oldname, newname string) error {
+	return f.SymlinkError
+}
+
+type FakeStater struct {
+	StatFileInfo     os.FileInfo
+	StatError        error
+	IsNotExistResult bool
+}
+
+func (f *FakeStater) Stat(name string) (os.FileInfo, error) {
+	return f.StatFileInfo, f.StatError
+}
+
+func (f *FakeStater) IsNotExist(err error) bool {
+	return f.IsNotExistResult
+}
+
+type MkdirSymlinker interface {
+	DirMaker
+	Symlinker
+}
+
 // Actual implementation of interface using os package.
 
 type OsDirMaker struct {
@@ -57,8 +73,7 @@ func (f *OsDirMaker) MkdirAll(path string, perm os.FileMode) error {
 	return os.MkdirAll(path, perm)
 }
 
-type OsSymlinker struct {
-}
+type OsSymlinker struct{}
 
 func (f *OsSymlinker) Symlink(oldname, newname string) error {
 	return os.Symlink(oldname, newname)
@@ -67,4 +82,14 @@ func (f *OsSymlinker) Symlink(oldname, newname string) error {
 type OsMkdirSymlinker struct {
 	*OsDirMaker
 	*OsSymlinker
+}
+
+type OsStater struct{}
+
+func (s *OsStater) Stat(name string) (os.FileInfo, error) {
+	return os.Stat(name)
+}
+
+func (s *OsStater) IsNotExist(err error) bool {
+	return os.IsNotExist(err)
 }
